@@ -20,11 +20,13 @@
         "optimize-autoloader": true
     },
     "require":{
-        "php": ">=5.4.0",
-        "teknoo/curl-request": "~1.0,>=1.0.1",
+        "php": ">=7.0.0",
+        "psr/http-message":"^1.0.0",
         "typo3/class-alias-loader":"~1.0"
     },
     "require-dev": {
+        "guzzlehttp/guzzle":"^6.0.0",
+        "guzzlehttp/psr7":"^1.0.0",
         "phpunit/phpunit": "~4.0",
         "phpunit/phpunit-mock-objects": "~2.3",
         "phpunit/phpunit-selenium": "~1.4",
